:art: Move the settingsActivity file back - missclick
